documentation + fixed missing subcommand error
