+++ hadoop-hdds/tools/src/main/java/org/apache/hadoop/hdds/scm/cli/datanode/DatanodeCommands.java
@@ -19,11 +19,13 @@ package org.apache.hadoop.hdds.scm.cli.datanode;
 
 import java.util.concurrent.Callable;
 
+import org.apache.hadoop.hdds.cli.GenericCli;
 import org.apache.hadoop.hdds.cli.HddsVersionProvider;
-import org.apache.hadoop.hdds.cli.MissingSubcommandException;
 import org.apache.hadoop.hdds.scm.cli.container.WithScmClient;
 
 import picocli.CommandLine;
+import picocli.CommandLine.Model.CommandSpec;
+import picocli.CommandLine.Spec;
 
 /**
  * Subcommand for datanode related operations.
@@ -38,6 +40,9 @@ import picocli.CommandLine;
     })
 public class DatanodeCommands implements Callable<Void> {
 
+  @Spec
+  private CommandSpec spec;
+
   @CommandLine.ParentCommand
   private WithScmClient parent;
 
@@ -47,7 +52,7 @@ public class DatanodeCommands implements Callable<Void> {
 
   @Override
   public Void call() throws Exception {
-    throw new MissingSubcommandException(
-        new CommandLine(new DatanodeCommands()));
+    GenericCli.missingSubcommand(spec);
+    return null;
   }
 }
