+++ hadoop-hdds/tools/src/main/java/org/apache/hadoop/hdds/scm/cli/pipeline/PipelineCommands.java
@@ -19,13 +19,14 @@ package org.apache.hadoop.hdds.scm.cli.pipeline;
 
 import java.util.concurrent.Callable;
 
+import org.apache.hadoop.hdds.cli.GenericCli;
 import org.apache.hadoop.hdds.cli.HddsVersionProvider;
-import org.apache.hadoop.hdds.cli.MissingSubcommandException;
 import org.apache.hadoop.hdds.scm.cli.container.WithScmClient;
 
-import picocli.CommandLine;
 import picocli.CommandLine.Command;
+import picocli.CommandLine.Model.CommandSpec;
 import picocli.CommandLine.ParentCommand;
+import picocli.CommandLine.Spec;
 
 /**
  * Subcommand to group pipeline related operations.
@@ -44,6 +45,9 @@ import picocli.CommandLine.ParentCommand;
     })
 public class PipelineCommands implements Callable<Void> {
 
+  @Spec
+  private CommandSpec spec;
+
   @ParentCommand
   private WithScmClient parent;
 
@@ -53,7 +57,7 @@ public class PipelineCommands implements Callable<Void> {
 
   @Override
   public Void call() throws Exception {
-    throw new MissingSubcommandException(
-        new CommandLine(new PipelineCommands()));
+    GenericCli.missingSubcommand(spec);
+    return null;
   }
 }
